Input Controller and Meteor
Added a meteor object with an IsTrigger box collider (also added one to the rocketship sprite gameobject). When it collides with another sprite object, it hill changes its color from white to red and back to white if else any other color. Also added a general script to be used that is atteched to the main camera. when the escape key is pressed, this will exit the game application once in a created build