added audio explosion sound efffect when projectiles collide with asteroids
added audio sources to meteros and and audio listener to pawn prefabs. created a separate audioplayer component that handles acutally playing the audio. The main audio player function works by getting set information. of which being what audio source is going to be played. that actually is played by calling a function of the audio source's conencted gameobject scripting, which uses Reflection to match the AudioClip variable names with a passed in string that checks for the desired variable, then passes the matching variable's value back to them be played by the audio player